+++ Return.h
@@ -14,8 +14,8 @@ public:
 
   ~Return();
 
-  void add(string dr, int amount, string returner, string bc, string dB);
-  void updateReturn(int amount, string newDateReturned);
+  void add(string dr, double amount, string returner, string bc, string dB);
+  void updateReturn(double amount, string newDateReturned);
   int getReturn();
   string getDatePurchased();
   string getCompany();
