+++ returnto.h
@@ -16,8 +16,8 @@ public:
 
   ~ReturnTo();
 
-  void tui(string dr, int amount, string bc, string dB);
-  void updateReturn(int amount, string c);
+  void tui(string dr, double amount, string bc, string dB);
+  void updateReturn(double amount, string c);
   int getReturn();
   string getDatePurchased();
   string getCompany();
